pyproject.toml
[tool.poetry]
name = "online platform electronics shop"
version = "0.1.0"
description = ""
authors = ["Mikhail Polyakov PyCharm <mikhapol@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
coverage = "^7.3.2"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
